+++ .gitignore
@@ -2,6 +2,7 @@
 __pycache__/
 db.sqlite3
 media
+migrations/
 
 # Backup files # 
 *.bak
