+++ scotlandyard/asgi.py
@@ -13,7 +13,7 @@ from channels.auth import AuthMiddlewareStack
 from channels.routing import ProtocolTypeRouter, URLRouter
 from django.core.asgi import get_asgi_application
 
-from scotlandyardgame.routing import websocket_urlpatterns
+from scotlandyardserver.routing import websocket_urlpatterns
 
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'scotlandyard.settings')
