+++ scotlandyard/urls.py
@@ -17,7 +17,6 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path('', include('scotlandyardgame.urls')),
-    # path('scotlandyard/', include('scotlandyardgame.urls')),
+    path('', include('scotlandyardserver.urls')),
     path('admin/', admin.site.urls),
 ]
