--- scotlandyardgame/templates/scotlandyardgame/index.html
@@ -1,136 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="shortcut icon" href="img/favicon.png" />
-    {% comment %}
-    <meta
-      name="theme-color"
-      media="(prefers-color-scheme: light)"
-      content="cyan"
-    />
-
-    <script src="/sw.js"></script>
-    {% endcomment %} {% comment %}
-    <link rel="manifest" href="/manifest.json" />
-    {% endcomment %}
-
-    <title>Scotland Yard</title>
-
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
-      crossorigin="anonymous"
-    />
-    <link
-      rel="stylesheet"
-      href="https://code.getmdl.io/1.3.0/material.indigo-pink.min.css"
-    />
-    <script defer src="https://code.getmdl.io/1.3.0/material.min.js"></script>
-
-    <link rel="stylesheet" href="/styles/main.css" />
-    <link rel="stylesheet" href="/styles/index.css" />
-
-    <link
-      rel="prefetch"
-      as="style"
-      href="https://fonts.googleapis.com/icon?family=Material+Icons"
-    />
-    <link
-      rel="prefetch"
-      href="https://fonts.gstatic.com/s/materialicons/v121/flUhRq6tzZclQEJ-Vdg-IuiaDsNc.woff2"
-      as="font"
-      type="font/woff2"
-    />
-    <link rel="prefetch" href="/js/anime.es.js" as="script" />
-    <script src="/js/index.js" type="module"></script>
-  </head>
-
-  <body>
-    <div id="preload" class="overlay fade-out"></div>
-    <div id="mobile-overlay" class="overlay">
-      <h3 class="display-3">Scotland Yard is not currently available on mobile.</h3>
-    </div>
-    <div class="layout-transparent mdl-layout mdl-js-layout" id="main">
-      <header class="mdl-layout__header mdl-layout__header--transparent">
-        <div class="mdl-layout__header-row">
-          <span class="mdl-layout-title display-1" id="logo">scotlandyard</span>
-          <div class="mdl-layout-spacer"></div>
-          <nav class="mdl-navigation">
-            <a class="mdl-navigation__link" href="help.html" style="color: #111"
-              >Help</a
-            >
-          </nav>
-        </div>
-      </header>
-
-      <main class="mdl-layout__content" id="main-title">
-        <div class="container">
-          <div class="title">
-            <h1>Scotland Yard</h1>
-            <form
-              action="/{{game_id}}"
-              method="post"
-              style="margin: auto"
-              id="start"
-            >
-              {% csrf_token %} {% if error is not None %}
-              <span id="error">{{ error }}</span>
-              {% endif %} {% if is_joining %}
-              <span id="room">
-                Joining room: <strong>{{ game_id }}</strong>.
-                <a class="link-light" href="/" id="create-link"
-                  >Create your own room.</a
-                ><br /><br />
-              </span>
-              {% endif %}
-              <div class="row g-2">
-                <div class="col-9">
-                  <input
-                    class="form-control"
-                    type="text"
-                    name="player_name"
-                    id="player_name"
-                    placeholder="Name"
-                    pattern="^\w+$"
-                    title="letters, numbers and underscores"
-                    required
-                  />
-                  <input
-                    type="hidden"
-                    id="game_id"
-                    name="game_id"
-                    value="{{ game_id }}"
-                  />
-                </div>
-                <div class="col-3">
-                  <button id="start-button" class="btn btn-primary">
-                    Start
-                  </button>
-                </div>
-              </div>
-            </form>
-          </div>
-        </div>
-      </main>
-      <footer class="footer">
-        <div class="container" id="version">
-          <span class="text-muted">{{version}}</span>
-        </div>
-      </footer>
-    </div>
-
-    <script
-      src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
-      integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
-      crossorigin="anonymous"
-    ></script>
-    <script
-      src="https://code.jquery.com/jquery-3.6.0.min.js"
-      integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4="
-      crossorigin="anonymous"
-    ></script>
-  </body>
-</html>
