--- scotlandyardgame/templates/scotlandyardgame/lobby.html
@@ -1,173 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="shortcut icon" href="img/favicon.png" />
-
-    <link
-      rel="stylesheet"
-      href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
-      integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
-      crossorigin="anonymous"
-    />
-    <link
-      rel="preload"
-      href="https://fonts.googleapis.com/icon?family=Material+Icons"
-      as="style"
-    />
-    <link
-      rel="stylesheet"
-      href="https://code.getmdl.io/1.3.0/material.indigo-pink.min.css"
-    />
-
-    <script defer src="https://code.getmdl.io/1.3.0/material.min.js"></script>
-    <title>Scotland Yard</title>
-    <link rel="stylesheet" href="/styles/main.css" />
-    <link rel="stylesheet" href="/styles/lobby.css" />
-    <link
-      rel="stylesheet"
-      href="https://fonts.googleapis.com/icon?family=Material+Icons"
-    />
-    <title>Lobby | Scotland Yard</title>
-  </head>
-
-  <body>
-    <div id="preload" class="overlay">
-      <div class="preload-svg-wrap">
-        <svg class="previewsvg" viewbox="0 0 100 100">
-          <g id="-preload-playerbody" transform="translate(27, 0)">
-            <path
-              class="playerheadpath"
-              id="preload-head"
-              stroke-width="3"
-              fill="#0000"
-            />
-            <path
-              class="playerbodypath"
-              id="preload-body"
-              stroke-width="3"
-              fill="#0000"
-              d="M 0 100 L 25 50 L 50 100 L 12 100"
-            />
-          </g>
-        </svg>
-      </div>
-    </div>
-
-    <div class="layout-transparent mdl-layout mdl-js-layout" id="layout">
-      <header
-        class="mdl-layout__header mdl-layout__header--transparent"
-        id="header"
-      >
-        <div class="mdl-layout__header-row">
-          <span class="mdl-layout-title display-1" id="logo">scotlandyard</span>
-        </div>
-      </header>
-
-      <main class="mdl-layout__content">
-        <div class="container" id="main">
-          <div id="players"></div>
-          <div id="buttons">
-            <button
-              id="colorButton"
-              class="button mdl-button mdl-js-button mdl-button--raised mdl-js-ripple-effect mdl-button--colored"
-              data-toggle="modal"
-              data-target="#colorModal"
-            >
-              <span class="material-icons">palette</span> Color
-            </button>
-            <button
-              class="button mdl-button mdl-js-button mdl-button--raised mdl-js-ripple-effect mdl-button--colored"
-              id="copy-link"
-            >
-              <span class="material-icons">content_copy</span> Copy Invite
-            </button>
-            <button
-              class="button mdl-button mdl-js-button mdl-button--raised mdl-js-ripple-effect"
-              id="start"
-            >
-              <span class="material-icons">play_arrow</span> Start
-            </button>
-          </div>
-          <div id="leave">
-            <a class="btn btn-danger" id="leave_button">means Leave da</a>
-          </div>
-        </div>
-      </main>
-    </div>
-
-    <span id="link" style="display: none"></span>
-
-    <div class="modal fade" id="colorModal" tabindex="-1" role="dialog">
-      <div class="modal-dialog modal-dialog-centered" role="document">
-        <div class="modal-content">
-          <div class="modal-header">
-            <h6 class="modal-title">select color</h6>
-            <button type="button" class="close" data-dismiss="modal">
-              &times;
-            </button>
-          </div>
-          <div class="modal-body">
-            <div class="row">
-              <div id="preview" class="col-6">
-                <svg class="previewsvg" viewbox="0 0 100 100">
-                  <g id="playerbody" transform="translate(27, 0)">
-                    <path
-                      class="playerheadpath"
-                      id="preview-head"
-                      stroke-width="3"
-                      fill="#0000"
-                    />
-                    <path
-                      class="playerbodypath"
-                      id="preview-body"
-                      stroke-width="3"
-                      fill="#0000"
-                      d="M 0 100 L 25 50 L 50 100 L 12 100"
-                    />
-                  </g>
-                </svg>
-              </div>
-              <div id="colorOptions" class="col-6">
-                <div id="colorList" class="list-group"></div>
-              </div>
-            </div>
-          </div>
-          <div class="modal-footer">
-            <button
-              id="saveColor"
-              type="button"
-              class="btn btn-primary"
-              data-dismiss="modal"
-            >
-              Close
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <div class="overlay" id="start-overlay">
-      <div id="overlay-text">starting game</div>
-    </div>
-
-    <script
-      src="https://code.jquery.com/jquery-3.6.0.min.js"
-      integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4="
-      crossorigin="anonymous"
-    ></script>
-    <script
-      src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"
-      integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q"
-      crossorigin="anonymous"
-    ></script>
-    <script
-      src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"
-      integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl"
-      crossorigin="anonymous"
-    ></script>
-    <script src="/js/lobby.js" type="module"></script>
-  </body>
-</html>
