+++ scotlandyardserver/__init__.py
@@ -1,8 +1,7 @@
-from .__version__ import __version__
-
-__name__ = "scotlandyardgame"
+__version__ = "0.1.1a"
 __author__ = "Romir Kulshrestha, Bhavvya Sarna"
 __license__ = "MIT"
 __copyright__ = "Copyright (c) Romir Kulshrestha, Bhavvya Sarna 2022"
+__url__ = "scotlandyard.imagifight.in"
 
 VERSION = __version__
