+++ scotlandyardserver/engine/ForceDirectedRenderer.py
