+++ scotlandyardserver/engine/mapdata.py
@@ -1,0 +1,78 @@
+MAP_DATA = [
+    # index: [taxi, bus, underground, special]
+    [[8, 9], [46, 58], [46]],
+    [[20, 10]],
+    [[11, 12, 4], [22, 23]],
+    [[3, 13]],
+    [[15, 16]],
+    [[29, 7]],
+    [[6, 17], [42]],
+    [[1, 18, 19]],
+    [[1, 19, 20]],
+    [[2, 11, 21, 34]],
+    [[3, 10, 22]],
+    [[3, 23]],
+    [[4, 14, 23, 24], [14, 23, 52], [46, 67, 89]],
+    [[13, 15, 25], [13, 15]],
+    [[5, 14, 16, 26, 28], [14, 29, 41]],
+    [[5, 16, 28, 29]],
+    [[7, 29, 30]],
+    [[8, 31, 43]],
+    [[8, 9, 32]],
+    [[2, 9, 33]],
+    [[10, 33]],
+    [[11, 23, 34, 35], [3, 23, 34, 65]],
+    [[12, 13, 22, 37], [3, 22, 23, 67]],
+    [[23, 37, 38]],
+    [[14, 38, 39]],
+    [[15, 27, 39]],
+    [[26, 28, 40]],
+    [[15, 16, 27, 41]],
+    [[6, 16, 17, 41, 42], [15, 41, 42, 55]],
+    [[17, 42]],
+    [[18, 43, 44]],
+    [[19, 33, 44, 45]],
+    [[20, 21, 32, 46]],
+    [[10, 22, 47, 48], [22, 46, 63]],
+    [[22, 36, 48, 65]],
+    [[35, 37, 49]],
+    [[23, 24, 36, 50]],
+    [[24, 25, 50, 51]],
+    [[25, 26, 51, 52]],
+    [[27, 41, 52, 53]],
+    [[28, 29, 40, 54], [15, 29, 52, 87]],
+    [[29, 30, 56, 72], [7, 29, 72]],
+    [[18, 31, 57]],
+    [[31, 32, 58]],
+    [[32, 46, 58, 59, 60]],
+    [[33, 45, 47, 61], [1, 34, 58, 78], [1, 13, 74, 79]],
+    [[34, 46, 62]],
+    [[34, 35, 62, 63]],
+    [[36, 50, 66]],
+    [[37, 38, 49]],
+    [[38, 39, 52, 67, 68]],
+    [[39, 40, 51, 69], [13, 41, 67, 86]],
+    [[40, 54, 69]],
+    [[41, 53, 55, 70]],
+    [[29, 54, 71], [29, 89]],
+    [[42, 91]],
+    [[43, 58, 73]],
+    [[44, 45, 57, 59, 74], [1, 46, 74, 77]],
+    [[45, 58, 75, 76]],
+    [[45, 61, 76]],
+    [[46, 60, 62, 76, 78]],
+    [[47, 48, 61, 79]],
+    [[48, 64, 79, 80], [34, 65, 79, 100]],
+    [[63, 65, 81]],
+    [[35, 64, 66, 82], [22, 63, 67, 82]],
+    [[49, 65, 67, 82]],
+    [[51, 66, 68, 84], [23, 52, 65, 82, 102], [13, 79, 89, 111]],
+    [[51, 67, 69, 85]],
+    [[52, 68, 53, 86]],
+    [[54, 71, 87]],
+    [[55, 70, 72, 89]],
+    [[42, 71, 90, 91], [42, 105, 107]],
+    [[57, 74, 92]],
+    [[58, 73, 75, 92], [58, 94], [46]],
+    [[58, 59, 74, 94]],
+]
