+++ scotlandyardserver/patterns.py
@@ -1,0 +1,4 @@
+import re
+
+UUID_RE = r"[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}"
+NAME_RE_PATTERN = re.compile(r"^\w+$")
