+++ scotlandyardserver/templates/scotlandyardclient/game.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="shortcut icon" href="img/favicon.png" />
+
+    <title>Document</title>
+    <link rel="stylesheet" href="styles/main.css" />
+    <link rel="stylesheet" href="styles/game.css" />
+
+    <style>
+      .turn {
+        color: var(--color-{{color}});
+      }
+    </style>
+  </head>
+  <body>
+    <main id="board">
+      <canvas id="canvas"></canvas>
+      <div id="ui">
+        <code id="log"> </code>
+        <br />
+        <code
+          >>>>&nbsp;<span contenteditable="true" id="ws-command"></span
+        ></code>
+      </div>
+    </main>
+    <script
+      src="https://code.jquery.com/jquery-3.6.0.min.js"
+      integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4="
+      crossorigin="anonymous"
+    ></script>
+    <script src="/js/game.js" type="module"></script>
+  </body>
+</html>
