+++ scotlandyardserver/templates/scotlandyardclient/game_old.html
