+++ scotlandyardserver/ws/GameConsumer.py
